optimize optimize optimize

+++ src/noise_generator.cpp
@@ -36,7 +36,7 @@ void NoiseGenerator::generateNextNoises()
   noise_cond_.notify_all();
 }
 
-std::tuple<xt::xtensor<float, 2> &, xt::xtensor<float, 2> &, xt::xtensor<float, 2> &> 
+std::tuple<xt::xtensor<float, 2> &, xt::xtensor<float, 2> &, xt::xtensor<float, 2> &>
 NoiseGenerator::getNoises()
 {
   std::unique_lock<std::mutex> guard(noise_lock_);
@@ -51,9 +51,9 @@ void NoiseGenerator::reset(mppi::models::OptimizerSettings & settings, bool is_h
   // Recompute the noises on reset, initialization, and fallback
   {
     std::unique_lock<std::mutex> guard(noise_lock_);
-    noises_vx_ = xt::zeros<double>({settings_.batch_size, settings_.time_steps});
-    noises_vy_ = xt::zeros<double>({settings_.batch_size, settings_.time_steps});
-    noises_wz_ = xt::zeros<double>({settings_.batch_size, settings_.time_steps});
+    noises_vx_ = xt::zeros<float>({settings_.batch_size, settings_.time_steps});
+    noises_vy_ = xt::zeros<float>({settings_.batch_size, settings_.time_steps});
+    noises_wz_ = xt::zeros<float>({settings_.batch_size, settings_.time_steps});
     ready_ = true;
   }
   noise_cond_.notify_all();
@@ -63,7 +63,7 @@ void NoiseGenerator::noiseThread()
 {
   do {
     std::unique_lock<std::mutex> guard(noise_lock_);
-    noise_cond_.wait(guard, [this](){return ready_;});
+    noise_cond_.wait(guard, [this]() {return ready_;});
     ready_ = false;
     generateNoisedControls();
   } while (active_);
@@ -73,10 +73,10 @@ void NoiseGenerator::generateNoisedControls()
 {
   auto & s = settings_;
 
-  noises_vx_ = xt::random::randn<double>({s.batch_size, s.time_steps}, 0.0, s.sampling_std.vx);
-  noises_wz_ = xt::random::randn<double>({s.batch_size, s.time_steps}, 0.0, s.sampling_std.wz);
+  noises_vx_ = xt::random::randn<float>({s.batch_size, s.time_steps}, 0.0, s.sampling_std.vx);
+  noises_wz_ = xt::random::randn<float>({s.batch_size, s.time_steps}, 0.0, s.sampling_std.wz);
   if (is_holonomic_) {
-    noises_vy_ = xt::random::randn<double>({s.batch_size, s.time_steps}, 0.0, s.sampling_std.vy);
+    noises_vy_ = xt::random::randn<float>({s.batch_size, s.time_steps}, 0.0, s.sampling_std.vy);
   }
 }
